+++ Dockerfile
@@ -19,7 +19,7 @@ ADD https://raw.githubusercontent.com/asam-ev/ASAM-ODS-Interfaces/main/ods_exter
 RUN python3 -m grpc_tools.protoc -I. --python_out=. ods.proto
 RUN python3 -m grpc_tools.protoc -I. --python_out=. --grpc_python_out=. ods_external_data.proto
 # Copy plugin implementation
-COPY __init__.py exd_api_server.py external_data_reader.py ./
+COPY __init__.py exd_api_server.py external_data_reader.py external_data_file.py ./
 USER appuser
 # Start server
 CMD [ "python3", "exd_api_server.py"]
